requirements.txt
cffi
cryptography
pika
PyJWT
